# openapi/openapi_core.yaml
openapi: 3.0.3
info:
  title: Tatvacare CRM API
  description: Core API for Tatvacare HealthtechCRM system
  version: 1.0.0
servers:
  - url: '{baseUrl}'
    variables:
      baseUrl:
        default: 'https://api.tatvacare.com'
        description: API base URL
paths:
...
  /esign/webhook:
    post:
      summary: E-Sign Webhook
      operationId: handleESignWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                envelopeId:
                  type: string
                status:
                  type: string
                  enum: [sent, delivered, signed, completed, declined, voided]
      responses:
        '200':
          description: Webhook processed successfully
...